target works

+++ Assets/Scripts/CookingLaser.cs
@@ -45,7 +45,7 @@ public class CookingLaser : MonoBehaviour
             Target target = hit.transform.GetComponent<Target>();
             if(target != null)
             {
-                target.TakeDamage(damage);
+                target.TakeDamage((int)damage);
             }
 
             if(hit.rigidbody != null)
